Create coordinate_generator file and test
